+++ crates/core-executor/src/datafusion/logical_analyzer/mod.rs
@@ -4,7 +4,7 @@ use std::sync::Arc;
 
 pub mod cast_analyzer;
 pub mod iceberg_types_analyzer;
-pub mod union_schema_cast_analyzer;
+pub mod union_schema_analyzer;
 
 #[must_use]
 pub fn analyzer_rules(
@@ -16,7 +16,7 @@ pub fn analyzer_rules(
         Arc::new(cast_analyzer::CastAnalyzer::new(session_params)),
         // Must be registered after CastAnalyzer because it introduces function calls
         // that can change the schema
-        Arc::new(union_schema_cast_analyzer::UnionSchemaAnalyzer::new()),
+        Arc::new(union_schema_analyzer::UnionSchemaAnalyzer::new()),
     ];
     base_rules.extend(rules);
     base_rules
