+++ crates/core-executor/src/datafusion/logical_analyzer/union_schema_analyzer.rs
