+++ crates/core-executor/src/tests/sql/logical_analyzer/mod.rs
@@ -1,2 +1,2 @@
 pub mod casting;
-mod union_schema_cast_analyzer;
+mod union_schema_analyzer;
